Crate bevy_macro_utils

A collection of helper types and functions for working on macros within the Bevy ecosystem.

Modules

fq_std

This module contains unit structs that should be used inside quote! and spanned_quote! using the variable interpolation syntax in place of their equivalent structs and traits present in std.

Structs

BevyManifest

The path to the Cargo.toml file for the Bevy project.

ResultSifter

Helper struct used to process an iterator of Result<Vec<T>, syn::Error>, combining errors into one along the way.

Symbol

A single named value, representable as a string.

Functions

derive_label

Derive a label trait

ensure_no_collision

Finds an identifier that will not conflict with the specified set of tokens.

get_lit_bool

Get a literal boolean from the provided expression as a bool.

get_lit_str

Get a literal string from the provided expression.

get_struct_fields

Get the fields of a data structure if that structure is a struct with named fields; otherwise, return a compile error that points to the site of the macro invocation.

terminated_parser

Returns a syn::parse::Parser which parses a stream of zero or more occurrences of T separated by punctuation of type P, with optional trailing punctuation.