bevy_macro_utils
Function ensure_no_collision

pub fn ensure_no_collision(value: Ident, haystack: TokenStream) -> Ident

Finds an identifier that will not conflict with the specified set of tokens.

If the identifier is present in haystack, extra characters will be added to it until it no longer conflicts with anything.

Note that the returned identifier can still conflict in niche cases, such as if an identifier in haystack is hidden behind an un-expanded macro.